#2158: split shared_cos_lease/mod.rs into cohesive submodules
The shared CoS lease module file crossed the ~2,000-LOC modularity
threshold (2,121 LOC). Split it into same-module submodules following the
converged plan section 5.4 and the established Rust sibling-submodule
pattern. This is pure code-motion of the hot-path CoS shaper: bodies move
verbatim, no logic change, no afxdp-facing API change.

New layout (all well under threshold):

  mod.rs       92  thin shell — mod/use wiring + re-exports + test mod
  backlog.rs  210  SharedCoSExactBacklog + residual-surplus token bucket
  vtime.rs    179  cross-worker MQFQ V_min floor (PaddedVtimeSlot,
                   SharedCoSQueueVtimeFloor)
  epoch.rs    536  #1229 Phase-6 v8 epoch state: V8State,
                   SharedCoSEpochState, PackedEpochGrant, the v8
                   enums/consts, equal-flow suppress state
  lease.rs   1325  SharedCoS{Queue,Root}Lease token bucket + the v8
                   per-worker fair-share acquire path + free helpers

backlog.rs and vtime.rs are self-contained and need no visibility change.

Visibility widenings (the only non-move edits). A private item is
visible only in its defining module and that module's descendants; a
sibling submodule is not a descendant. Before the split these epoch/lease
types lived in mod.rs (the parent), so every sibling saw their
inherent-private fields for free. Moving the definitions into child
submodules makes those fields private-to-the-child, so each item a
sibling reaches across the new file boundary widens from inherent-private
to pub(super) — the same minimal widening the already-extracted
rotate_epoch_v8.rs / publish_equal_flow_epoch_v8.rs (#1588) documented.
No new pub / pub(crate); no cross-crate surface change. Each widening is
justified inline at its definition:

  epoch.rs
    - consts EPOCH_DURATION_NS, MAX_ROTATION_LAG_EPOCHS,
      STALL_THRESHOLD_EPOCHS, CARRY_MAX_EPOCHS, CARRY_DRAIN_MAX_EPOCHS,
      MAX_SEQ_SPINS, MAX_ROLLBACK_RETRIES, EQUAL_FLOW_VALID_STREAK_REQUIRED,
      EQUAL_FLOW_MIN_WORKER_UTIL_NUM/_DEN — read by lease.rs /
      rotate_epoch_v8.rs / publish_equal_flow_epoch_v8.rs.
    - PackedEpochGrant (struct + its tuple field .0 + pack/unpack/new)
      and PaddedAtomicU64 (struct + field .0) — loaded/CAS'd by siblings.
    - SharedCoSEpochState (struct + every field a sibling reads) and its
      ::new ctor.
    - V8State (struct + all fields — each is reached by some sibling).
    - V8RotationScratch (struct + fields + ::new) — built by lease.rs,
      fields read by rotate_epoch_v8.rs.
    - V8EqualFlowSuppressState (struct + all fields + new/fail_open/
      enforce_epoch/disable_for_epoch) — reached by lease.rs and the
      publish/rotate siblings.
    - V8EqualFlowFailOpenReason::from_u32/as_str — called by lease.rs.
    (The v8/cause/fail-open enums keep their original pub(in crate::afxdp)
    visibility; they are re-exported by types/mod.rs.)

  lease.rs
    - SharedCoSQueueLease.config and .v8 — read by rotate_epoch_v8.rs.
    - SharedCoSLeaseConfig (struct, named via the pub(super) config
      field) + its rate_bytes field (read by rotate_epoch_v8.rs).
    - test-only reaches restored to their pre-split inherent access:
      SharedCoSQueueLease.state, SharedCoSRootLease.config/.state,
      SharedCoSLeaseConfig.burst_bytes, SharedCoSLeaseState (+ credits/
      last_refill_ns), equal_flow_cap_v8, and the free fns
      pack/unpack_shared_cos_lease_credits, refill_shared_cos_lease_state,
      record_equal_flow_active_sample.

Hot-path codegen: userspace-dp builds with LTO off and codegen-units 16,
so a cross-file move can drop the implicit single-translation-unit
inlining the helper had. try_bump_outstanding (called from the acquire_v8
primary + surplus loops) gains an #[inline] hint to preserve it. The
other moved hot helpers (bump_epoch_event, record_equal_flow_active_sample,
worker_grant_bump, tag_checked_rollback, the publish/PackedEpochGrant
pack/unpack) already carried #[inline]/#[inline(always)] and keep it. The
large functions (acquire_v8 ~300 LOC, the rotation) are too big to inline
regardless, so their move is codegen-neutral.

The v8_carry_field_is_reader_private guard test allowlist is updated to
follow the SharedCoSEpochState definition from mod.rs to epoch.rs. The
rotation-private invariant is unchanged: the field's only reader/writer
is still the rotation winner in rotate_epoch_v8.rs.

Validation: release build clean (warning count unchanged vs master). Full
Rust suite matches master — the shared_cos_lease module is 88/88 and all
cos+lease named tests are 508/508, both green across 5 consecutive runs;
the suite's 3 fails (concurrent_recovery, wg reconcile_peers_snapshot,
tx_latency_hist skew) are pre-existing concurrency flakes verified failing
on the unmodified master baseline.
